+++ components/mailbox/src/Mailbox.svelte
@@ -56,6 +56,8 @@
   export let query_string: string; // Allowed query parameter list https://developer.nylas.com/docs/api/#get/threads
   export let show_star: boolean;
   export let show_thread_checkbox: boolean;
+  export let show_reply: boolean;
+  export let show_reply_all: boolean;
 
   const defaultValueMap: Partial<MailboxProperties> = {
     actions_bar: [],
@@ -63,6 +65,8 @@
     query_string: "in=inbox",
     show_star: false,
     show_thread_checkbox: true,
+    show_reply: false,
+    show_reply_all: false,
   };
 
   let manifest: Partial<MailboxProperties> = {};
@@ -682,6 +686,8 @@
           {you}
           show_star={_this.show_star}
           click_action="mailbox"
+          show_reply={_this.show_reply}
+          show_reply_all={_this.show_reply_all}
           on:messageClicked={messageClicked}
           on:threadStarred={threadStarred}
           on:returnToMailbox={returnToMailbox}
@@ -809,6 +815,8 @@
                     {you}
                     show_star={_this.show_star}
                     click_action="mailbox"
+                    show_reply={_this.show_reply}
+                    show_reply_all={_this.show_reply_all}
                     on:threadClicked={threadClicked}
                     on:messageClicked={messageClicked}
                     on:threadStarred={threadStarred}